Initial

+++ src/vapordmods/vapordmods.py
@@ -8,6 +8,8 @@ import json
 import zipfile
 from src.api import thunderstore, nexusmods
 
+logger = logging.getLogger(__name__)
+
 
 class vapordmods:
     __CFG_FILENAME = 'vapordmods.yml'
@@ -22,6 +24,10 @@ class vapordmods:
         self.mods_info = None
         self.thunderstore_dir = os.path.join(mods_dir, self.__MANIFESTS_DIR, self.__THUNDERSTORE_NAME)
         self.nexusmods_dir = os.path.join(mods_dir, self.__MANIFESTS_DIR, self.__NEXUSMODS_NAME)
+        self.mods_dir = {
+            self.__THUNDERSTORE_NAME: self.thunderstore_dir,
+            self.__NEXUSMODS_NAME: self.nexusmods_dir
+        }
 
         if os.path.exists(self.cfg_filename):
             if os.path.isfile(self.cfg_filename) and os.access(self.cfg_filename, os.R_OK):
@@ -86,30 +92,40 @@ class vapordmods:
                                 installed_version = currentmanifest['version_number']
                             if modversion == installed_version:
                                 data = {'currentversion': modversion, 'latestversion': 'current_match_requirement',
-                                        'needupdate': False, 'download_link': 'not_required', 'newmanifest': None}
+                                        'needupdate': False, 'download_link': 'not_required',
+                                        'mods_dir': 'not_required', 'newmanifest': None}
                         else:
-                            r = await globals()[provider].get_update(app, modname, modversion, api_key)
-                            if r:
-                                latestversion = r['data']['version']
+                            params = {}
+                            if provider == self.__THUNDERSTORE_NAME:
+                                params = {
+                                    'namespace': app,
+                                    'name': modname,
+                                    'version': modversion
+                                }
+
+                            modclass = getattr(globals()[provider], provider)
+                            modobject = modclass()
+                            req = await modobject.get_update(**params)
+                            if req == 0:
+                                r = modobject.return_data()
+                                latestversion = r['version']
                                 download_link = None
                                 if currentmanifest:
                                     currentversion = currentmanifest['version']
-                                    if currentmanifest['version'] == r['data']['version']:
+                                    if currentmanifest['version'] == r['version']:
                                         needupdate = False
-                                        download_link = 'not_required'
                                         newmanifest = None
                                     else:
                                         needupdate = True
-                                        ddownload_link = r['data']['download_url']
-                                        newmanifest = r['data']['manifest']
+                                        newmanifest = r['manifest']
                                 else:
                                     currentversion = 'not_installed'
                                     needupdate = True
-                                    download_link = r['data']['download_url']
-                                    newmanifest = r['data']['manifest']
+                                    newmanifest = r['manifest']
 
                                 data = {'currentversion': currentversion, 'latestversion': latestversion,
-                                        'needupdate': needupdate, 'download_link': download_link, 'newmanifest': newmanifest}
+                                        'needupdate': needupdate, 'download_link': r['download_url'],
+                                        'mods_dir': self.mods_dir[provider], 'newmanifest': newmanifest}
                             else:
                                 logging.error("Impossible de trouver l'info pour ...")
 
@@ -127,23 +143,25 @@ class vapordmods:
             try:
                 destination = os.path.join(self.mods_dir, mods_name)
                 file.extractall(destination)
+                return 0
             except EOFError as er:
                 raise er
 
-    async def make_request(self, session, url, mods_name):
+    async def make_request(self, session, mods_name, update_manifest):
         try:
-            resp = await session.request(method="GET", url=url)
+            resp = await session.request(method="GET", url=update_manifest['download_link'])
         except Exception as ex:
             print(ex)
             return
 
         if resp.status == 200:
-            filename = os.path.join(self.mods_dir, resp.url.name)
+            filename = os.path.join(update_manifest['mods_dir'], resp.url.name)
             async with aiofiles.open(filename, 'wb') as f:
                 await f.write(await resp.read())
 
-            async with aiofiles.open(filename, 'r') as f:
-                await asyncio.get_running_loop().run_in_executor(None, self.extract_mods, resp.url.name, mods_name)
+            if await asyncio.get_running_loop().run_in_executor(None, self.extract_mods, resp.url.name, mods_name) == 0:
+                async with aiofiles.open(manifest['namespace'] + '-' + manifest['name'], 'w') as file_manifest:
+                    await file_manifest.write(json.dumps(manifest, indent=4))
                 os.remove(filename)
 
     async def update_mods(self):
@@ -159,7 +177,7 @@ class vapordmods:
                                 for app_k, app_v in provider_v.items():
                                     for mods_k, mods_v in app_v.items():
                                         if mods_v['needupdate']:
-                                            tasks.append(self.make_request(session, mods_v['download_link'], mods_k))
+                                            tasks.append(self.make_request(session, mods_k, mods_v))
 
                             if len(tasks) > 0:
                                 await asyncio.gather(*tasks)
